+++ src/Option.ts
@@ -13,6 +13,7 @@ import * as B from "fp-ts/boolean"
 import { invert as invertBool } from "./Boolean"
 import { pureIf as _pureIf } from "./Alternative"
 import { toMonoid as _toMonoid } from "./Monoid"
+import { Lazy } from "./Lazy"
 
 /**
  * Unwrap the value from within an `Option`, throwing if `None`.
@@ -135,19 +136,21 @@ export const memptyUnless: (x: boolean) => <A>(m: Option<A>) => Option<A> =
   flow(invertBool, memptyWhen)
 
 /**
- * Conditionally lifts a value to `Some` or returns `None`.
+ * Conditionally lifts a value to `Some` or returns `None`. The lazy value is
+ * evaluated only if the condition passes.
  *
  * @example
+ * import { constant } from 'fp-ts/function';
  * import { pureIf } from 'fp-ts-std/Option';
  * import { Predicate } from 'fp-ts/Predicate';
  *
  * const person = { name: 'Hodor', age: 40 };
  * const isMagicNumber: Predicate<number> = n => n === 42;
  *
- * const mname = pureIf(isMagicNumber(person.age))(person.name);
+ * const mname = pureIf(isMagicNumber(person.age))(constant(person.name));
  *
  * @since 0.13.0
  */
-export const pureIf: (x: boolean) => <A>(y: A) => Option<A> = _pureIf(
+export const pureIf: (x: boolean) => <A>(y: Lazy<A>) => Option<A> = _pureIf(
   O.Alternative,
 )